# perf: lazy load SVG icons