--- a/frappe/website/utils.py
+++ b/frappe/website/utils.py
@@ -563,10 +563,18 @@ def set_content_type(response, data, path):
 
 def add_preload_for_bundled_assets(response):
 	links = [f"<{css}>; rel=preload; as=style" for css in frappe.local.preload_assets["style"]]
-
 	links.extend(f"<{js}>; rel=preload; as=script" for js in frappe.local.preload_assets["script"])
+	links.extend(_preload_svg_headers())
+
 	if links:
 		response.headers["Link"] = ",".join(links)
+
+
+def _preload_svg_headers():
+	include_icons = frappe.get_hooks().get("app_include_icons", [])
+
+	version = frappe.cache.get_value("metadata_version")
+	return [f"</assets/{svg}?v={version}>; rel=preload; as=fetch" for svg in include_icons]
 
 
 @lru_cache